* Add group.

+++ module/story/view/blocksibling.html.php
@@ -5,6 +5,7 @@ js::set('relieved',    $lang->story->relieved);
 js::set('relievedTip', $lang->story->relievedTip);
 js::set('cancel',      $lang->cancel);
 $canViewLinkStory = common::hasPriv('story', 'view');
+$canRelieved      = common::hasPriv('story', 'relieved');
 
 foreach($siblings as $id => $sibling)
 {
@@ -21,7 +22,7 @@ foreach($siblings as $sibling)
     echo "<li title='$title' class='sibling'>";
     echo "<span class='label label-outline label-badge'>{$branch}</span> ";
     echo ($canViewLinkStory ? html::a($this->createLink('story', 'view', "id=$id", '', true), "$title", '', "class='iframe' data-width='80%'") : "$id $title");
-    echo "<a class='unlink hide' data-id='$id' data-toggle='popover'><i class='icon icon-unlink btn-info'></i></a>";
+    if($canRelieved) echo "<a class='unlink hide' data-id='$id' data-toggle='popover'><i class='icon icon-unlink btn-info'></i></a>";
     echo "</li>";
 }
 ?>
